Changes to Collisions and Camera, mob rotation included.

+++ ComGraphics/ComGraphics/Source/MobBossOne.h
@@ -23,7 +23,7 @@ public:
 private:
     bool kenaWhack;
     
-    int animtimer;
+    float animtimer;
